fix comments on PR first step
