+++ src/libraries/icubmod/CMakeLists.txt
@@ -55,7 +55,7 @@ yarp_begin_plugin_library(icubmod QUIET)
   add_subdirectory(imuFilter)
   add_subdirectory(embObjPSC)
   add_subdirectory(embObjPOS)
-  add_subdirectory(embObjCanBatterysensor)
+  add_subdirectory(embObjBattery)
 if (ICUB_ICUBINTERFACE_EXPERIMENTAL)
   add_subdirectory(imu3DM_GX3)
 endif()
