+++ src/libraries/icubmod/embObjBattery/CMakeLists.txt
@@ -4,9 +4,9 @@
 # This software may be modified and distributed under the terms of the
 # BSD-3-Clause license. See the accompanying LICENSE file for details.
 
-yarp_prepare_plugin(embObjCanBatterysensor CATEGORY device
-                                 TYPE yarp::dev::embObjCanBatterysensor
-                                 INCLUDE embObjCanBatterysensor.h
+yarp_prepare_plugin(embObjBattery CATEGORY device
+                                 TYPE yarp::dev::embObjBattery
+                                 INCLUDE embObjBattery.h
                                  EXTRA_CONFIG WRAPPER=analogServer)
 
 IF (NOT SKIP_embObjCanBatterysensor)
@@ -17,23 +17,23 @@ IF (NOT SKIP_embObjCanBatterysensor)
 
   INCLUDE_DIRECTORIES(${CMAKE_CURRENT_SOURCE_DIR})
 
-  yarp_add_plugin(embObjCanBatterysensor embObjCanBatterysensor.cpp embObjCanBatterysensor.h)
-  TARGET_LINK_LIBRARIES(embObjCanBatterysensor ethResources)
-  icub_export_plugin(embObjCanBatterysensor)
+  yarp_add_plugin(embObjBattery embObjBattery.cpp embObjBattery.h)
+  TARGET_LINK_LIBRARIES(embObjBattery ethResources)
+  icub_export_plugin(embObjBattery)
 
-  yarp_install(TARGETS embObjCanBatterysensor
+  yarp_install(TARGETS embObjBattery
         COMPONENT Runtime
         LIBRARY DESTINATION ${ICUB_DYNAMIC_PLUGINS_INSTALL_DIR}
         ARCHIVE DESTINATION ${ICUB_STATIC_PLUGINS_INSTALL_DIR}
         YARP_INI DESTINATION ${ICUB_PLUGIN_MANIFESTS_INSTALL_DIR})
 
   if (BUILD_TESTING)
-   add_library(embObjCanBatterysensorUT STATIC embObjCanBatterysensor.cpp embObjCanBatterysensor.h)    
-   target_link_libraries(embObjCanBatterysensorUT ethResources
+   add_library(embObjBatteryUT STATIC embObjBattery.cpp embObjBattery.h)    
+   target_link_libraries(embObjBatteryUT ethResources
                                     YARP::YARP_dev
                                     icub_firmware_shared::embobj)
                                         
-   target_include_directories(embObjCanBatterysensorUT PUBLIC "$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>")
+   target_include_directories(embObjBatteryUT PUBLIC "$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>")
   endif()
 
 ENDIF (NOT SKIP_embObjCanBatterysensor)
